+++ Kernel/System/DynamicField/Backend.pm
@@ -2193,6 +2193,8 @@ The following functions should be only used if the dynamic field has IsStatsCond
         Name               => 'DynamicField_' . $DynamicFieldConfig->{Label},
         Element            => 'DynamicField_' . $DynamicFieldConfig->{Name},
         TranslatableValues => 1,
+        TimePeriodFormat   => 'DateInputFormat',
+        Block              => 'InputField',              # or 'MultiselectField' or 'Time'
     };
 
 =cut